+++ test/lib/Elastica/Test/Base.php
@@ -7,12 +7,6 @@ use Elastica\Index;
 
 class Base extends \PHPUnit_Framework_TestCase
 {
-    protected function es20($message = '')
-    {
-        parent::setUp();
-        $this->markTestIncomplete($message."\nTest skipped because of current incompatibility with ES 2.0");
-    }
-
     /**
      * @param array    $params   Additional configuration params. Host and Port are already set
      * @param callback $callback