+++ test/lib/Elastica/Test/BulkTest.php
@@ -676,7 +676,7 @@ class BulkTest extends BaseTest
 
         $endMemory = memory_get_usage();
 
-        $this->es20('Failed asserting that 2.2414096568375803 is less than 1.3.');
+        $this->markTestIncomplete('Failed asserting that 2.2414096568375803 is less than 1.3.');
         $this->assertLessThan(1.3, $endMemory / $startMemory);
     }
 }